requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.2.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
filelock==3.15.4
frozenlist==1.4.0
fsspec==2024.6.1
gmpy2==2.1.2
google-auth==2.29.0
google-auth-oauthlib==0.5.2
grpcio==1.62.2
idna==3.7
importlib-metadata==7.0.1
Jinja2==3.1.4
lightning-utilities==0.11.6
Markdown==3.4.1
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pip==24.2
protobuf==4.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.2.1
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
setuptools==72.1.0
sympy==1.12
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.11.0
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
yarl==1.9.3
zipp==3.17.0
